+++ docs/release-notes.md
@@ -12,7 +12,9 @@ hide:
 
 - ManyToManyField `create_through_model` method allows now the keyword only argument `replace_related_field`.
 - `add_to_registry` has now an additional keyword-only argument `replace_related_field_m2m` for extra controlling the `create_through_model` registration logic.
+- `add_to_registry` has now an additional keyword-only argument `on_conflict` for controlling what happens when a same named model already exists.
 - Passing a tuple or list of types to `replace_related_field` is now allowed.
+- Add `through_registry` to ManyToMany.
 
 ### Changed
 
@@ -22,6 +24,8 @@ hide:
 ### Fixed
 
 - Copying registries is working now.
+- Fix deleting (clearing cache) of BaseForeignKey target.
+- Creating two models with the same name did lead to silent replacements.
 
 ## 0.24.2